refactor tree utils
